v3

+++ ft_bzero.c
@@ -17,7 +17,7 @@ void	ft_bzero(void *s, size_t n)
 	size_t		i;
 
 	i = 0;
-	while ((size_t)i < n)
+	while (i < n)
 	{
 		((char *)s)[i] = 0;
 		i++;
